LU-5577 obdclass: cl_lock_nr_mutexed() return type to size_t
Change the type accordant usage.

+++ lustre/obdclass/cl_lock.c
@@ -143,7 +143,7 @@ static void cl_lock_trace0(int level, const struct lu_env *env,
 {
         struct cl_object_header *h = cl_object_header(lock->cll_descr.cld_obj);
         CDEBUG(level, "%s: %p@(%d %p %d %d %d %d %d %lx)"
-                      "(%p/%d/%d) at %s():%d\n",
+                      "(%p/%d/%zu) at %s():%d\n",
 	       prefix, lock, atomic_read(&lock->cll_ref),
                lock->cll_guarder, lock->cll_depth,
                lock->cll_state, lock->cll_error, lock->cll_holds,
@@ -774,11 +774,11 @@ EXPORT_SYMBOL(cl_lock_is_mutexed);
 /**
  * Returns number of cl_lock mutices held by the current thread (environment).
  */
-int cl_lock_nr_mutexed(const struct lu_env *env)
+size_t cl_lock_nr_mutexed(const struct lu_env *env)
 {
-        struct cl_thread_info *info;
-        int i;
-        int locked;
+	struct cl_thread_info *info;
+	size_t i;
+	size_t locked;
 
         /*
          * NOTE: if summation across all nesting levels (currently 2) proves
